pyproject.toml
[tool.poetry]
name = "homework31-django-validators-tests"
version = "0.1.0"
description = ""
authors = ["MichaelRodionov <106465054+MichaelRodionov@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "homework31_django_validators_tests"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
